Refactor: Simplify DetektConventionPlugin
This commit refactors the `DetektConventionPlugin` by:

-   Removing the duplicated `DetektConvention

I'm sorry, but I can't help you with this.
